+++ services/leases/service.go
@@ -18,13 +18,11 @@ package leases
 
 import (
 	"context"
-	"errors"
 
 	api "github.com/containerd/containerd/api/services/leases/v1"
 	"github.com/containerd/containerd/errdefs"
 	"github.com/containerd/containerd/leases"
 	"github.com/containerd/containerd/plugin"
-	"github.com/containerd/containerd/services"
 	ptypes "github.com/gogo/protobuf/types"
 	"google.golang.org/grpc"
 )
@@ -34,18 +32,10 @@ func init() {
 		Type: plugin.GRPCPlugin,
 		ID:   "leases",
 		Requires: []plugin.Type{
-			plugin.ServicePlugin,
+			plugin.LeasePlugin,
 		},
 		InitFn: func(ic *plugin.InitContext) (interface{}, error) {
-			plugins, err := ic.GetByType(plugin.ServicePlugin)
-			if err != nil {
-				return nil, err
-			}
-			p, ok := plugins[services.LeasesService]
-			if !ok {
-				return nil, errors.New("leases service not found")
-			}
-			i, err := p.Instance()
+			i, err := ic.GetByID(plugin.LeasePlugin, "manager")
 			if err != nil {
 				return nil, err
 			}
