+++ services/services.go
@@ -29,8 +29,6 @@ const (
 	TasksService = "tasks-service"
 	// NamespacesService is id of namespaces service.
 	NamespacesService = "namespaces-service"
-	// LeasesService is id of leases service.
-	LeasesService = "leases-service"
 	// DiffService is id of diff service.
 	DiffService = "diff-service"
 	// IntrospectionService is the id of introspection service
